backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
llama-index==0.10.33  
pydantic-settings==2.1.0
python-multipart==0.0.6
llama-index-llms-together
llama-index-embeddings-huggingface
psycopg2-binary
